chore: rename dir to align with transport later
